% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CpiAdjustment.R
\name{applyCpiAdjustmentSql}
\alias{applyCpiAdjustmentSql}
\title{Apply CPI Adjustment in SQL}
\usage{
applyCpiAdjustmentSql(costColumn, yearColumn, cpiTable, defaultFactor = 1)
}
\arguments{
\item{costColumn}{Name of the cost column}

\item{yearColumn}{Name of the year column}

\item{cpiTable}{Name of the CPI adjustment table}

\item{defaultFactor}{Default adjustment factor if year not found}
}
\value{
SQL expression for adjusted cost
}
\description{
Generates SQL snippet to apply CPI adjustment to costs
}
\examples{
# Generate SQL for CPI adjustment
sql <- applyCpiAdjustmentSql("total_cost", "YEAR(incurred_date)", "cpi_adjustment")
}
