% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataQuality.R
\name{assertCdmConforms}
\alias{assertCdmConforms}
\title{Assert CDM Conforms to CostUtilization Assumptions}
\usage{
assertCdmConforms(connection, cdmDatabaseSchema)
}
\arguments{
\item{connection}{A connection object from `DatabaseConnector`.}

\item{cdmDatabaseSchema}{The schema holding the OMOP CDM data.}
}
\description{
This function checks if the target CDM has a COST table and if the
visit-centric linkage assumption holds. It will stop execution if
the assumptions are not met.
}
\keyword{internal}
