% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CpiAdjustment.R
\name{calculateCpiAdjustmentFactor}
\alias{calculateCpiAdjustmentFactor}
\title{Calculate CPI Adjustment Factor}
\usage{
calculateCpiAdjustmentFactor(fromYear, toYear, cpiData)
}
\arguments{
\item{fromYear}{The year of the original cost}

\item{toYear}{The target year for standardization}

\item{cpiData}{A tibble with year and cpi columns}
}
\value{
Numeric adjustment factor
}
\description{
Calculates the adjustment factor to convert costs from one year to another
}
\examples{
cpiData <- loadCpiData()
# Adjust from 2015 to 2023 dollars
factor <- calculateCpiAdjustmentFactor(2015, 2023, cpiData)
}
