% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conceptSetCostSettings.R
\name{combineCostCovariateSettings}
\alias{combineCostCovariateSettings}
\title{Combine multiple cost covariate settings}
\usage{
combineCostCovariateSettings(...)
}
\arguments{
\item{...}{Multiple costCovariateSettings objects to combine}
}
\value{
A combined costCovariateSettings object
}
\description{
Combines multiple cost covariate settings objects into a single settings object.
}
