% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CostCovariateData.R
\docType{class}
\name{CostCovariateData-class}
\alias{CostCovariateData-class}
\title{The CostCovariateData class}
\description{
A class for storing cost and utilization covariate data extracted from the database.
This object uses the Andromeda package to store data on disk, which allows it
to handle data sets that are much larger than the available RAM.
}
\section{Fields}{

\describe{
\item{\code{andromeda}}{An Andromeda object containing the covariate data.}

\item{\code{metaData}}{A list of metadata about the object.}
}}

\seealso{
\code{\link{getDbCostData}}
}
