% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CostUtilization-package.R
\docType{package}
\name{CostUtilization-package}
\alias{CostUtilization}
\alias{CostUtilization-package}
\title{CostUtilization: Healthcare Cost and Utilization Analysis for OMOP CDM v5.5}
\description{
The CostUtilization package provides a comprehensive framework for analyzing
healthcare costs and resource utilization using OMOP Common Data Model (CDM)
databases. This version is specifically designed for \strong{CDM v5.5} with enhanced
support for the new long-format COST table structure.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OHDSI/CostUtilization}
  \item Report bugs at \url{https://github.com/OHDSI/CostUtilization/issues}
}

}
\author{
\strong{Maintainer}: Gowtham Rao \email{gowthamrao@gmail.com}

Other contributors:
\itemize{
  \item Observational Health Data Science and Informatics [copyright holder]
}

}
\keyword{internal}
