% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureExtractionIntegration.R
\docType{class}
\name{CovariateData-class}
\alias{CovariateData-class}
\alias{CovariateData}
\alias{show,CovariateData-method}
\alias{summary,CovariateData-method}
\title{The CovariateData Class}
\usage{
\S4method{show}{CovariateData}(object)

\S4method{summary}{CovariateData}(object)
}
\arguments{
\item{object}{An object of class \code{CovariateData}.}
}
\description{
A custom S4 class that inherits from \code{Andromeda::Andromeda}. It is designed to
store covariate data from a \code{CostUtilization} analysis in a format compatible
with the OHDSI FeatureExtraction standard.
}
\seealso{
\code{\link[=createCostCovariateData]{createCostCovariateData()}}
}
