% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsFormat.R
\name{.assembleCovariateData}
\alias{.assembleCovariateData}
\title{Assemble CovariateData Object}
\usage{
.assembleCovariateData(
  covariates,
  covariateRef,
  analysisRef,
  timeRef = NULL,
  metaData
)
}
\arguments{
\item{covariates}{Covariates tibble}

\item{covariateRef}{Covariate reference tibble}

\item{analysisRef}{Analysis reference tibble}

\item{timeRef}{Time reference tibble (optional)}

\item{metaData}{Metadata list}
}
\value{
CovariateData object
}
\description{
Assemble CovariateData Object
}
\keyword{internal}
