% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsFormat.R
\name{.createAnalysisName}
\alias{.createAnalysisName}
\title{Create Analysis Name}
\usage{
.createAnalysisName(costOfCareSettings)
}
\arguments{
\item{costOfCareSettings}{Settings object}
}
\value{
Descriptive analysis name
}
\description{
Create Analysis Name
}
\keyword{internal}
