% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsFormat.R
\name{.createAnalysisRef}
\alias{.createAnalysisRef}
\title{Create Analysis Reference Table}
\usage{
.createAnalysisRef(analysisId, costOfCareSettings, databaseId)
}
\arguments{
\item{analysisId}{Analysis ID}

\item{costOfCareSettings}{Settings object}

\item{databaseId}{Database identifier}
}
\value{
Tibble of analysis references
}
\description{
Create Analysis Reference Table
}
\keyword{internal}
