% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsFormat.R
\name{.createCovariateMapping}
\alias{.createCovariateMapping}
\title{Create Covariate Mapping for Aggregated Results}
\usage{
.createCovariateMapping(resultsData, analysisId, costOfCareSettings)
}
\arguments{
\item{resultsData}{Tibble of aggregated results}

\item{analysisId}{Analysis ID for covariate numbering}

\item{costOfCareSettings}{Settings object}
}
\value{
Tibble with covariate mappings
}
\description{
Create Covariate Mapping for Aggregated Results
}
\keyword{internal}
