% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsFormat.R
\name{.createCovariateRef}
\alias{.createCovariateRef}
\title{Create Covariate Reference Table}
\usage{
.createCovariateRef(covariateMapping, costOfCareSettings)
}
\arguments{
\item{covariateMapping}{Covariate mapping tibble}

\item{costOfCareSettings}{Settings object}
}
\value{
Tibble of covariate references
}
\description{
Create Covariate Reference Table
}
\keyword{internal}
