% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsFormat.R
\name{.createMetaData}
\alias{.createMetaData}
\title{Create Comprehensive Metadata}
\usage{
.createMetaData(
  costOfCareSettings,
  cohortId,
  databaseId,
  analysisId,
  resultFormat
)
}
\arguments{
\item{costOfCareSettings}{Settings object}

\item{cohortId}{Cohort ID}

\item{databaseId}{Database identifier}

\item{analysisId}{Analysis ID}

\item{resultFormat}{Result format type}
}
\value{
List of metadata
}
\description{
Create Comprehensive Metadata
}
\keyword{internal}
