% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsFormat.R
\name{.createPersonLevelCovariateMapping}
\alias{.createPersonLevelCovariateMapping}
\title{Create Covariate Mapping for Person-Level Results}
\usage{
.createPersonLevelCovariateMapping(resultsData, analysisId, costOfCareSettings)
}
\arguments{
\item{resultsData}{Tibble of person-level results}

\item{analysisId}{Analysis ID for covariate numbering}

\item{costOfCareSettings}{Settings object}
}
\value{
Tibble with covariate mappings
}
\description{
Create Covariate Mapping for Person-Level Results
}
\keyword{internal}
