% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsFormat.R
\name{.formatCovariateName}
\alias{.formatCovariateName}
\title{Format Covariate Name}
\usage{
.formatCovariateName(metricName, costOfCareSettings)
}
\arguments{
\item{metricName}{Raw metric name}

\item{costOfCareSettings}{Settings object for context}
}
\value{
Formatted covariate name
}
\description{
Format Covariate Name
}
\keyword{internal}
