% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsFormat.R
\name{.generateCovariatesFromAggregated}
\alias{.generateCovariatesFromAggregated}
\title{Generate Covariates from Aggregated Results}
\usage{
.generateCovariatesFromAggregated(resultsData, covariateMapping, cohortId)
}
\arguments{
\item{resultsData}{Aggregated results tibble}

\item{covariateMapping}{Covariate mapping tibble}

\item{cohortId}{Cohort ID}
}
\value{
Tibble of covariates
}
\description{
Generate Covariates from Aggregated Results
}
\keyword{internal}
