% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsFormat.R
\name{.generateCovariatesFromPersonLevel}
\alias{.generateCovariatesFromPersonLevel}
\title{BUG FIX: This function was refactored to be robust and correct.}
\usage{
.generateCovariatesFromPersonLevel(resultsData, covariateMapping, cohortId)
}
\description{
BUG FIX: This function was refactored to be robust and correct.
}
