% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OhdsiFeatureExtractionIntegration.R
\name{.generateReferenceTables}
\alias{.generateReferenceTables}
\title{Internal helper to generate reference tables from the settings object}
\usage{
.generateReferenceTables(settings)
}
\description{
Internal helper to generate reference tables from the settings object
}
\keyword{internal}
