% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildAnalysis.R
\name{executeSqlPlan}
\alias{executeSqlPlan}
\title{Execute SQL Plan for Cost Analysis}
\usage{
executeSqlPlan(
  connection,
  params,
  targetDialect,
  tempEmulationSchema,
  verbose = TRUE
)
}
\description{
Execute SQL Plan for Cost Analysis
}
