% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CostCovariateData.R
\name{filterByAnalysisId}
\alias{filterByAnalysisId}
\title{Filter CostCovariateData by analysis ID}
\usage{
filterByAnalysisId(costCovariateData, analysisIds)
}
\arguments{
\item{costCovariateData}{CostCovariateData object}

\item{analysisIds}{Vector of analysis IDs to keep}
}
\value{
Filtered CostCovariateData object
}
\description{
Filter CostCovariateData by analysis ID
}
