% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CostCovariateData.R
\name{filterByTimeWindow}
\alias{filterByTimeWindow}
\title{Filter CostCovariateData by time window}
\usage{
filterByTimeWindow(costCovariateData, startDay, endDay)
}
\arguments{
\item{costCovariateData}{CostCovariateData object}

\item{startDay}{Start day of window}

\item{endDay}{End day of window}
}
\value{
Filtered CostCovariateData object
}
\description{
Filter CostCovariateData by time window
}
