% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildAnalysis.R
\name{generateCostAnalysisSql}
\alias{generateCostAnalysisSql}
\title{Generate Dynamic SQL for Cost Analysis}
\usage{
generateCostAnalysisSql(settings, cohortIds = -1, .route)
}
\arguments{
\item{settings}{Cost utilization settings object}

\item{cohortIds}{Vector of cohort IDs to analyze}
}
\value{
Character string containing the complete SQL
}
\description{
Generate Dynamic SQL for Cost Analysis
}
