% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDbCostData.R
\name{generateCostCovariatesSql}
\alias{generateCostCovariatesSql}
\title{Generate SQL for cost covariates extraction}
\usage{
generateCostCovariatesSql(
  connection,
  cdmDatabaseSchema,
  cohortTable,
  cohortDatabaseSchema,
  cohortId,
  covariateSettings,
  aggregated
)
}
\arguments{
\item{connection}{A \code{DatabaseConnector} connection object.}

\item{cdmDatabaseSchema}{The schema holding the OMOP CDM data.}

\item{cohortTable}{The table containing the cohorts.}

\item{cohortDatabaseSchema}{The schema where the cohort table resides.}

\item{cohortId}{The ID of the cohort for which to extract data.}

\item{covariateSettings}{An object of type \code{costCovariateSettings}.}

\item{aggregated}{A logical value indicating whether to compute aggregated statistics
or extract person-level data.}
}
\value{
A list containing the rendered and translated SQL queries for covariates,
covariate reference, and analysis reference.
}
\description{
This internal function constructs the necessary SQL queries to extract cost and
utilization covariates. It creates and uploads a temporary table for temporal
windows to avoid complex string parsing in SQL, making the queries more robust
and efficient.
}
\keyword{internal}
