% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildAnalysis.R
\name{generateFilters}
\alias{generateFilters}
\title{Generate filter conditions based on settings}
\usage{
generateFilters(settings, domainConceptIds)
}
\description{
Generate filter conditions based on settings
}
