% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureExtractionIntegration.R
\name{generateReferenceTables}
\alias{generateReferenceTables}
\title{Internal helper to generate reference tables from settings}
\usage{
generateReferenceTables(costCovariateSettings)
}
\description{
Internal helper to generate reference tables from settings
}
\keyword{internal}
