% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureExtractionIntegration.R
\name{generateReferenceTablesFromUnifiedSettings}
\alias{generateReferenceTablesFromUnifiedSettings}
\title{Internal helper to generate reference tables from the UNIFIED settings object}
\usage{
generateReferenceTablesFromUnifiedSettings(covariateSettings)
}
\description{
Internal helper to generate reference tables from the UNIFIED settings object
}
\keyword{internal}
