% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildAnalysis.R
\name{generateWindowsSql}
\alias{generateWindowsSql}
\title{Generate SQL for time windows}
\usage{
generateWindowsSql(settings)
}
\arguments{
\item{settings}{A costUtilizationSettings object.}
}
\value{
A SQL string for creating a temporary table of time windows.
}
\description{
Generate SQL for time windows
}
