% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDbCostData.R
\name{getCdmVersion}
\alias{getCdmVersion}
\title{Get CDM version from database}
\usage{
getCdmVersion(connection, cdmDatabaseSchema)
}
\arguments{
\item{connection}{Database connection}

\item{cdmDatabaseSchema}{CDM database schema}
}
\value{
CDM version string
}
\description{
Get CDM version from database
}
\keyword{internal}
