% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{getCostCovariateData}
\alias{getCostCovariateData}
\title{Get Cost Covariate Data}
\usage{
getCostCovariateData(
  connection,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable,
  cohortId,
  costCovariateSettings
)
}
\arguments{
\item{connection}{A connection object from the `DatabaseConnector` package.}

\item{cdmDatabaseSchema}{The schema holding the OMOP CDM data.}

\item{cohortDatabaseSchema}{The schema where the cohort table is located.}

\item{cohortTable}{The name of the cohort table.}

\item{cohortId}{The ID of the cohort to analyze.}

\item{costCovariateSettings}{A settings object created by `createCostCovariateSettings`.}
}
\value{
A FeatureExtraction-compatible data frame with the computed covariates.
}
\description{
Computes the cost covariates for a specified cohort based on the provided settings.
}
