% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCostUtilData.R
\name{getCostUtilData}
\alias{getCostUtilData}
\title{Get Cost and Utilization Data From the Database}
\usage{
getCostUtilData(
  connection,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable,
  cohortIds,
  costUtilSettings,
  aggregated = FALSE,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema")
)
}
\arguments{
\item{connection}{A connection to the server containing the OMOP CDM instance.}

\item{cdmDatabaseSchema}{The name of the database schema that contains the OMOP CDM instance.}

\item{cohortDatabaseSchema}{The name of the database schema where the cohort table is located.}

\item{cohortTable}{The name of the table holding the cohort(s).}

\item{cohortIds}{A vector of \code{cohort_definition_id}s to retrieve covariates for.}

\item{costUtilSettings}{An S3 object of type \code{costUtilSettings} as created by \code{createCostUtilSettings()}.}

\item{aggregated}{A logical flag indicating whether to compute population-level aggregate statistics
(\code{TRUE}) or person-level data (\code{FALSE}).}

\item{tempEmulationSchema}{A schema with write privileges for emulating temp tables on platforms like Oracle.}
}
\value{
A \code{CovariateData} object, which is an \code{Andromeda} object containing the results.
}
\description{
Constructs covariates for cost and healthcare utilization for one or more cohorts. This is the main
workhorse function of the \code{CostUtilization} package.
}
\details{
This function constructs a set of covariates based on the provided settings. It returns results
as a standard \code{CovariateData} object compatible with the OHDSI ecosystem.
}
