% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsFormat.R
\name{getCovariateValues}
\alias{getCovariateValues}
\title{Get Covariate Values for Specific People}
\usage{
getCovariateValues(covariateData, rowIds)
}
\arguments{
\item{covariateData}{CovariateData object}

\item{rowIds}{Vector of row IDs (person IDs) to extract}
}
\value{
Tibble with covariate values
}
\description{
Get Covariate Values for Specific People
}
