% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCostCovariates.R
\name{getDbCostCovariateData}
\alias{getDbCostCovariateData}
\title{Get Cost and Utilization Covariates From the Database}
\usage{
getDbCostCovariateData(
  connection,
  oracleTempSchema = NULL,
  cdmDatabaseSchema,
  cohortTable,
  cohortIds,
  cdmVersion = "5",
  rowIdField = "subject_id",
  costCovariateSettings,
  aggregated = FALSE,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  targetDatabaseSchema = NULL,
  targetCovariateTable = NULL,
  targetCovariateRefTable = NULL,
  targetAnalysisRefTable = NULL
)
}
\arguments{
\item{connection}{A connection to the server containing the OMOP CDM instance.}

\item{oracleTempSchema}{DEPRECATED. Use \code{tempEmulationSchema} instead.}

\item{cdmDatabaseSchema}{The name of the database schema that contains the OMOP CDM instance.}

\item{cohortTable}{The name of the table holding the cohort(s).}

\item{cohortIds}{A vector of \code{cohort_definition_id}s to retrieve covariates for.}

\item{cdmVersion}{The version of the OMOP CDM. Currently, only "5" is supported.}

\item{rowIdField}{The name of the field in the cohort table that is to be used as the
\code{row_id} field. Crucial for person-level data.}

\item{costCovariateSettings}{An S3 object of type \code{costCovariateSettings} as created by \code{createCostCovariateSettings()}.}

\item{aggregated}{A logical flag indicating whether to compute population-level aggregate statistics
(\code{TRUE}) or person-level data (\code{FALSE}).}

\item{tempEmulationSchema}{A schema with write privileges for emulating temp tables.}

\item{targetDatabaseSchema}{(Optional) The name of the database schema where the resulting covariate tables
should be stored.}

\item{targetCovariateTable}{(Optional) The name of the table where the resulting covariates will
be stored. If not provided, results will be fetched to R.}

\item{targetCovariateRefTable}{(Optional) The name of the table where the covariate reference will be stored.}

\item{targetAnalysisRefTable}{(Optional) The name of the table where the analysis reference will be stored.}
}
\value{
If \code{targetCovariateTable} is not specified, a \code{CovariateData} object is returned.
Otherwise, nothing is returned.
}
\description{
Constructs covariates for cost and healthcare utilization for one or more cohorts.
}
\details{
This function constructs a set of covariates based on the provided settings. It can return the results
as a \code{CovariateData} object in memory, or write them to tables in the database.
}
