% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDbCostData.R
\name{getDbCostData}
\alias{getDbCostData}
\title{Get cost data from database}
\usage{
getDbCostData(
  connectionDetails,
  cdmDatabaseSchema,
  cohortTable = "cohort",
  cohortDatabaseSchema = cdmDatabaseSchema,
  cohortId,
  covariateSettings,
  aggregated = TRUE,
  minCharacterizationMean = 0.001,
  tempEmulationSchema = NULL
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.}

\item{cohortTable}{Name of the table with the cohort data.}

\item{cohortDatabaseSchema}{Schema where the cohort table resides. Default is cdmDatabaseSchema.}

\item{cohortId}{The cohort definition ID of the cohort for which to extract covariates.}

\item{covariateSettings}{An object of type \code{costCovariateSettings} as created using
the \code{costCovariateSettings} function.}

\item{aggregated}{Should aggregate statistics be computed? If TRUE, will return
population-level statistics. If FALSE, will return person-level data.}

\item{minCharacterizationMean}{Minimum mean value for characterization output.
Values below this will be censored.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support
temp tables. To emulate temp tables, provide a schema where temp tables can be created.}
}
\value{
An object of type \code{CostCovariateData} containing:
\itemize{
\item{covariates}{Cost covariates data}
\item{covariateRef}{Reference table with covariate definitions}
\item{analysisRef}{Reference table with analysis definitions}
\item{metaData}{Metadata about the extraction}
}
}
\description{
Extracts cost covariate data from a CDM database based on the provided settings.
This function handles the connection to the database, SQL generation, and data retrieval.
}
