% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OhdsiFeatureExtractionIntegration.R
\name{getDbCovariateData.costUtilSettings}
\alias{getDbCovariateData.costUtilSettings}
\title{Construct Cost and Utilization Covariates}
\usage{
\method{getDbCovariateData}{costUtilSettings}(
  connection,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable,
  cohortId = NULL,
  cohortIds,
  cdmVersion = "5",
  rowIdField = "subject_id",
  covariateSettings,
  aggregated = FALSE,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema")
)
}
\arguments{
\item{connection}{A connection to the server containing the schema.}

\item{cdmDatabaseSchema}{The name of the database schema that contains the OMOP CDM instance.}

\item{cohortDatabaseSchema}{The name of the database schema where the cohort table is located.}

\item{cohortTable}{Name of the table holding the cohort.}

\item{cohortId}{DEPRECATED. Use \code{cohortIds} instead.}

\item{cohortIds}{A vector of cohort definition IDs.}

\item{cdmVersion}{The version of the OMOP CDM.}

\item{rowIdField}{The name of the field in the cohort table that is to be used as the \code{row_id} field.}

\item{covariateSettings}{An S3 object of type \code{costUtilSettings}.}

\item{aggregated}{A logical flag indicating whether to compute aggregate statistics (\code{TRUE}) or person-level data (\code{FALSE}).}

\item{tempEmulationSchema}{A schema with write privileges for emulating temp tables.}
}
\value{
A \code{CovariateData} object.
}
\description{
This is an S3 method for the \code{FeatureExtraction::getDbCovariateData} generic.
It is called when the \code{covariateSettings} object is of type \code{costUtilSettings}.
}
\details{
This function is not intended to be called directly but is automatically dispatched
by \code{FeatureExtraction::getDbCovariateData}. It serves as a wrapper around the main
\code{getCostUtilData} function to ensure compatibility with the OHDSI framework.
}
