% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsFormat.R
\name{is.CovariateData}
\alias{is.CovariateData}
\title{Check if Object is CovariateData}
\usage{
is.CovariateData(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical
}
\description{
Check if Object is CovariateData
}
