% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costSettings.R
\name{isCostCovariateSettings}
\alias{isCostCovariateSettings}
\title{Check if settings are cost covariate settings}
\usage{
isCostCovariateSettings(object)
}
\arguments{
\item{object}{Object to check}
}
\value{
Logical indicating if object is costCovariateSettings
}
\description{
Check if settings are cost covariate settings
}
