% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureExtractionIntegration.R
\name{isTemporalCovariateData}
\alias{isTemporalCovariateData}
\title{Check if CovariateData is temporal}
\usage{
isTemporalCovariateData(x)
}
\arguments{
\item{x}{The \code{CovariateData} object to check.}
}
\value{
A logical value.
}
\description{
Checks if a \code{CovariateData} object contains temporal data (i.e., has a \code{timeId} column).
}
