% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CostCovariateData.R
\name{loadCostCovariateData}
\alias{loadCostCovariateData}
\title{Load CostCovariateData from a file}
\usage{
loadCostCovariateData(file)
}
\arguments{
\item{file}{The path to the file where the data was saved.}
}
\value{
A CostCovariateData object.
}
\description{
Load CostCovariateData from a file
}
