% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureExtractionIntegration.R
\name{loadCovariateData}
\alias{loadCovariateData}
\title{Load Covariate Data from a File}
\usage{
loadCovariateData(file)
}
\arguments{
\item{file}{The path to the file containing the data.}
}
\value{
An object of class \code{CovariateData}.
}
\description{
Loads a \code{CovariateData} object from a file.
}
