% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot.CostCovariateData}
\alias{plot.CostCovariateData}
\title{Plot CostCovariateData}
\usage{
\method{plot}{CostCovariateData}(x, type = "boxplot", covariateIds = NULL, ...)
}
\arguments{
\item{x}{A CostCovariateData object.}

\item{type}{The type of plot, either "boxplot" or "barchart".}

\item{covariateIds}{A vector of covariate IDs to include in the plot.}

\item{...}{Additional arguments (not used).}
}
\value{
A ggplot object.
}
\description{
Plot CostCovariateData
}
