% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot.CovariateData}
\alias{plot.CovariateData}
\title{Plot Covariate Data}
\usage{
\method{plot}{CovariateData}(x, type = "barchart", top_n = 20, title = NULL, ...)
}
\arguments{
\item{x}{A \code{CovariateData} object.}

\item{type}{The type of plot to create. Options are \code{"boxplot"} for person-level
data (\code{aggregated = FALSE}) or \code{"barchart"} for aggregated data
(\code{aggregated = TRUE}).}

\item{top_n}{The maximum number of covariates to plot, sorted by frequency (for boxplots)
or mean value (for barcharts). Defaults to 20.}

\item{title}{An optional title for the plot. If \code{NULL}, a default title is generated.}

\item{...}{Additional arguments (not used).}
}
\value{
A \code{ggplot} object.
}
\description{
Creates a plot of the covariate data, either as boxplots for person-level data
or bar charts for aggregated data. This is an S3 method for \code{CovariateData} objects.
}
