% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CostCovariateData.R
\name{print.CostCovariateData}
\alias{print.CostCovariateData}
\title{Print a CostCovariateData object}
\usage{
\method{print}{CostCovariateData}(x, ...)
}
\arguments{
\item{x}{A CostCovariateData object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object.
}
\description{
Print a CostCovariateData object
}
