% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costSettings.R
\name{print.costCovariateSettings}
\alias{print.costCovariateSettings}
\title{Print cost covariate settings}
\usage{
\method{print}{costCovariateSettings}(x, ...)
}
\arguments{
\item{x}{costCovariateSettings object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the input
}
\description{
Print cost covariate settings
}
