% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCostOfCareSettings.R
\name{print.CostOfCareSettings}
\alias{print.CostOfCareSettings}
\title{Print Cost of Care Settings}
\usage{
\method{print}{CostOfCareSettings}(x, ...)
}
\arguments{
\item{x}{A CostOfCareSettings object}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the input object
}
\description{
Print Cost of Care Settings
}
