% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CostUtilizationSettings.R
\name{print.costUtilizationSettings}
\alias{print.costUtilizationSettings}
\title{Print Cost and Utilization Settings}
\usage{
\method{print}{costUtilizationSettings}(x, ...)
}
\arguments{
\item{x}{An object of class \code{costUtilizationSettings}.}

\item{...}{For future extensions. Not used in this method.}
}
\description{
Provides a formatted, human-readable summary of the \code{costUtilizationSettings} object.
}
