% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CpiAdjustment.R
\name{print.CpiAdjustmentSettings}
\alias{print.CpiAdjustmentSettings}
\title{Print CPI Adjustment Settings}
\usage{
\method{print}{CpiAdjustmentSettings}(x, ...)
}
\arguments{
\item{x}{CPI adjustment settings object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print CPI Adjustment Settings
}
