% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureExtractionIntegration.R
\name{resolveConceptSet}
\alias{resolveConceptSet}
\title{Internal helper to resolve concept sets with descendants}
\usage{
resolveConceptSet(connection, cdmDatabaseSchema, conceptIds, addDescendants)
}
\description{
Internal helper to resolve concept sets with descendants
}
\keyword{internal}
