% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CostCovariateData.R
\name{saveCostCovariateData}
\alias{saveCostCovariateData}
\title{Save CostCovariateData to file}
\usage{
saveCostCovariateData(costCovariateData, file)
}
\arguments{
\item{costCovariateData}{CostCovariateData object}

\item{file}{Path to save file (RDS format)}
}
\description{
Save CostCovariateData to file
}
