% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureExtractionIntegration.R
\name{saveCovariateData}
\alias{saveCovariateData}
\title{Save Covariate Data to a File}
\usage{
saveCovariateData(covariateData, file)
}
\arguments{
\item{covariateData}{An object of type \code{CovariateData}.}

\item{file}{The path to the file where the data will be written.}
}
\value{
No return value, called for its side effect.
}
\description{
Saves a \code{CovariateData} object to a file.
}
