% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CpiAdjustment.R
\name{summarizeCpiAdjustedCosts}
\alias{summarizeCpiAdjustedCosts}
\title{Summarize CPI Adjusted Costs}
\usage{
summarizeCpiAdjustedCosts(
  costData,
  originalCostCol = "total_cost",
  adjustedCostCol = "adjusted_cost"
)
}
\arguments{
\item{costData}{A tibble with cost data}

\item{originalCostCol}{Name of original cost column}

\item{adjustedCostCol}{Name of adjusted cost column}
}
\value{
A tibble with summary statistics
}
\description{
Provides summary statistics for CPI-adjusted costs
}
