% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CostCovariateData.R
\name{summary.CostCovariateData}
\alias{summary.CostCovariateData}
\title{Summary of CostCovariateData}
\usage{
\method{summary}{CostCovariateData}(object, ...)
}
\arguments{
\item{object}{CostCovariateData object}

\item{...}{Additional arguments (ignored)}
}
\value{
Summary data frame
}
\description{
Summary of CostCovariateData
}
