% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CostCovariateData.R
\name{toAndromedaTable}
\alias{toAndromedaTable}
\title{Convert CostCovariateData to a standard Andromeda table}
\usage{
toAndromedaTable(costCovariateData)
}
\arguments{
\item{costCovariateData}{A CostCovariateData object.}
}
\value{
A \code{tbl_Andromeda} object. This is a pointer to the data on disk and can be
manipulated using \code{dplyr} verbs. Use \code{collect()} to bring the data into memory.
}
\description{
Convert CostCovariateData to a standard Andromeda table
}
