% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CostCovariateData.R
\name{toStandardCovariateData}
\alias{toStandardCovariateData}
\title{Convert CostCovariateData to standard covariate data format}
\usage{
toStandardCovariateData(costCovariateData)
}
\arguments{
\item{costCovariateData}{CostCovariateData object}
}
\value{
Data frame in standard covariate format
}
\description{
Convert CostCovariateData to standard covariate data format
}
