% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsFormat.R
\name{toWideFormat}
\alias{toWideFormat}
\title{Convert CovariateData to Wide Format}
\usage{
toWideFormat(covariateData)
}
\arguments{
\item{covariateData}{CovariateData object}
}
\value{
Tibble in wide format
}
\description{
Convert CovariateData to Wide Format
}
