% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CostUtilization-package.R
\docType{package}
\name{CostUtilization-package}
\alias{CostUtilization}
\alias{CostUtilization-package}
\title{CostUtilization: What the Package Does (One Line, Title Case)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a framework for executing standardized cost and healthcare utilization analyses on the OMOP Common Data Model. This package computes aggregated or person-level cost and utilization metrics across various domains (e.g., Drug, Visit) and temporal windows relative to a cohort. The primary output is a `CostCovariateData` object, which can be summarized, plotted, and used in downstream OHDSI study pipelines.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OHDSI/CostUtilization}
  \item Report bugs at \url{https://github.com/OHDSI/CostUtilization/issues}
}

}
\author{
\strong{Maintainer}: Gowtham Rao \email{gowthamrao@gmail.com}

Other contributors:
\itemize{
  \item Observational Health Data Science and Informatics [copyright holder]
}

}
\keyword{internal}
