% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDbCostData.R
\name{generateCostCovariatesSql}
\alias{generateCostCovariatesSql}
\title{Generate SQL for cost covariates extraction}
\usage{
generateCostCovariatesSql(
  connection,
  cdmDatabaseSchema,
  cohortTable,
  cohortDatabaseSchema,
  cohortId,
  covariateSettings,
  aggregated
)
}
\arguments{
\item{connection}{Database connection}

\item{cdmDatabaseSchema}{CDM database schema}

\item{cohortTable}{Cohort table name}

\item{cohortDatabaseSchema}{Cohort database schema}

\item{cohortId}{Cohort ID}

\item{covariateSettings}{Cost covariate settings}

\item{aggregated}{Whether to generate aggregated SQL}
}
\value{
List of SQL queries
}
\description{
Internal function to generate SQL queries for cost covariate extraction
}
\keyword{internal}
