% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCostUtilData.R
\name{getCostUtilData}
\alias{getCostUtilData}
\title{Get Cost and Utilization Data}
\usage{
getCostUtilData(
  connection = NULL,
  connectionDetails = NULL,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable,
  cohortIds,
  costUtilSettings,
  aggregated = FALSE
)
}
\arguments{
\item{connection}{An object of type \code{connection} as created using the
\code{DatabaseConnector::connect()} function.}

\item{connectionDetails}{An R object of type connectionDetails created using the
function createConnectionDetails in the DatabaseConnector package. Either
the connection or connectionDetails argument should be specified}

\item{cdmDatabaseSchema}{The name of the database schema that contains the OMOP CDM instance.
Requires read permissions to this schema.}

\item{cohortDatabaseSchema}{The name of the database schema that is the location where the
cohort data used for the analysis resides.}

\item{cohortTable}{The name of the table that contains the cohort data.}

\item{cohortIds}{A numeric vector of cohort definition IDs to be included in the analysis.}

\item{costUtilSettings}{An S3 object of type \code{costUtilSettings} created using the
\code{createCostUtilSettings()} function.}

\item{aggregated}{A boolean flag indicating whether to return aggregated summary
statistics (\code{TRUE}) or person-level data (\code{FALSE}).}
}
\value{
An R object of class \code{CovariateData}, which is an \code{Andromeda} object containing the
analysis results.
}
\description{
Executes the cost and utilization analysis against a cohort and returns the results
as a \code{CovariateData} object. This function uses a \code{costUtilSettings} object to define
the analysis parameters.
}
\details{
This function connects to the database, prepares the analysis settings, and runs the
main SQL query (\code{GetCostAndUtilization.sql}) to extract person-level data. If aggregation
is requested, it runs a second SQL query (\code{AggregateCovariates.sql}) to summarize the results.
The final output is structured as a standard OHDSI \code{CovariateData} object.
}
