% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDbCostData.R
\name{getDbCostData}
\alias{getDbCostData}
\title{Get cost data from database}
\usage{
getDbCostData(
  connectionDetails,
  cdmDatabaseSchema,
  cohortTable = "cohort",
  cohortDatabaseSchema = cdmDatabaseSchema,
  cohortId,
  covariateSettings,
  aggregated = TRUE,
  tempEmulationSchema = NULL
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails}.}

\item{cdmDatabaseSchema}{Schema name where your OMOP CDM data resides.}

\item{cohortTable}{Name of the table with the cohort data.}

\item{cohortDatabaseSchema}{Schema where the cohort table resides.}

\item{cohortId}{The cohort definition ID.}

\item{covariateSettings}{An object of type \code{costCovariateSettings}.}

\item{aggregated}{Should aggregate statistics be computed?}

\item{tempEmulationSchema}{A schema where temp tables can be created.}
}
\value{
A \code{CostCovariateData} object. This object is a pointer to data stored on disk.
Remember to close it using \code{Andromeda::close()} when you are finished.
}
\description{
Extracts cost covariate data from a CDM database based on the provided settings.
This function uses the Andromeda package to store data on disk, making it memory-efficient.
}
