% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCostUtilizationData.R
\name{getDbCostData}
\alias{getDbCostData}
\title{Get Cost and Utilization Data from the Database}
\usage{
getDbCostData(
  connectionDetails = NULL,
  connection = NULL,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable,
  cohortIds,
  costUtilizationSettings,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema")
)
}
\arguments{
\item{connectionDetails}{An R object of type \code{connectionDetails} created by the
\code{DatabaseConnector::createConnectionDetails} function.}

\item{cdmDatabaseSchema}{The name of the database schema that contains the OMOP CDM
instance.}

\item{cohortDatabaseSchema}{The name of the database schema that contains the cohort table.}

\item{cohortTable}{The name of the cohort table.}

\item{cohortIds}{A numeric vector of cohort definition IDs to include in the
analysis. If -1 - all cohorts in table will be used.}

\item{costUtilizationSettings}{An S3 object of class \code{costUtilizationSettings} created by the
\code{createCostUtilizationSettings()} function.}
}
\value{
A \code{tibble} containing the aggregated cost and HRU statistics.
}
\description{
This function connects to an OMOP CDM database, executes the cost and HRU
analysis based on the provided settings, and returns the results as a tibble.
}
\details{
This function uses \code{SqlRender} to parameterize a SQL script and \code{DatabaseConnector}
to execute it. All heavy computation is performed on the database server.

When a \code{conceptSetDefinition} is used, this function first executes a
preliminary query to identify the relevant domains, ensuring the main analysis
query is as efficient as possible by only scanning necessary event tables.
}
