% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCostUtilData.R
\name{getDbCostData}
\alias{getDbCostData}
\title{Get Cost and Utilization Data from the Database}
\usage{
getDbCostData(
  connectionDetails,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable,
  cohortIds,
  costUtilSettings
)
}
\arguments{
\item{connectionDetails}{An R object of type \code{connectionDetails} created by the
\code{DatabaseConnector::createConnectionDetails} function.}

\item{cdmDatabaseSchema}{The name of the database schema that contains the OMOP CDM
instance.}

\item{cohortDatabaseSchema}{The name of the database schema that contains the cohort table.}

\item{cohortTable}{The name of the cohort table.}

\item{cohortIds}{A numeric vector of cohort definition IDs to include in the
analysis.}

\item{costUtilSettings}{An S3 object of class \code{costUtilSettings} created by the
\code{costUtilizationSettings()} function.}
}
\value{
A \code{tibble} containing the aggregated cost and HRU statistics, with columns for
cohort ID, analysis category, analysis type (cost or hru), and descriptive statistics.
}
\description{
This function connects to an OMOP CDM database, executes the cost and HRU
analysis based on the provided settings, and returns the results as a tibble.
}
\details{
This function uses \code{SqlRender} to parameterize a SQL script and \code{DatabaseConnector}
to execute it. All heavy computation is performed on the database server. The
function handles the creation of temporary tables for intermediate steps,
including uploading cost standardization data if specified.
}
