% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureExtractionIntegration.R
\name{getDbCovariateData.costCovariateSettings}
\alias{getDbCovariateData.costCovariateSettings}
\title{Construct Cost and Utilization Covariates}
\usage{
getDbCovariateData.costCovariateSettings(
  connection,
  oracleTempSchema = NULL,
  cdmDatabaseSchema,
  cohortTable,
  cohortId,
  cdmVersion = "5",
  rowIdField,
  covariateSettings,
  aggregated = FALSE,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema")
)
}
\arguments{
\item{connection}{A connection to the server containing the schema.}

\item{oracleTempSchema}{DEPRECATED. Use \code{tempEmulationSchema} instead.}

\item{cdmDatabaseSchema}{The name of the database schema that contains the OMOP CDM instance.}

\item{cohortTable}{Name of the table holding the cohort.}

\item{cohortId}{The ID of the cohort for which to construct covariates.}

\item{cdmVersion}{The version of the OMOP CDM. Currently, only "5" is supported.}

\item{rowIdField}{The name of the field in the cohort table that is to be used as the
\code{row_id} field. Crucial for person-level data.}

\item{covariateSettings}{An S3 object of type \code{costCovariateSettings}.}

\item{aggregated}{A logical flag indicating whether to compute population-level aggregate statistics
(\code{TRUE}) or person-level data (\code{FALSE}).}

\item{tempEmulationSchema}{A schema with write privileges for emulating temp tables.}
}
\value{
A \code{CovariateData} object containing the constructed covariates.
If \code{aggregated = TRUE}, results are in \code{covariatesContinuous}.
If \code{aggregated = FALSE}, results are in \code{covariates}.
}
\description{
This is an S3 method for the \code{FeatureExtraction::getDbCovariateData} generic.
It is called when the \code{covariateSettings} object is of type \code{costCovariateSettings}.
This function constructs aggregated or person-level cost and healthcare utilization covariates.
}
\details{
This function is not intended to be called directly, but is automatically dispatched
by \code{FeatureExtraction::getDbCovariateData}. It generates covariates for the specified
cohort based on settings created by \code{createCostCovariateSettings}.
When \code{aggregated = TRUE}, it produces summary statistics.
When \code{aggregated = FALSE}, it produces person-level data, which requires a unique
\code{rowIdField} in the cohort table to distinguish cohort entries.
}
