% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EunomiaTestHelpers.R
\name{injectCostData}
\alias{injectCostData}
\title{Inject Synthetic Payer Plan and Wide-Format Cost Data into an OMOP CDM (Eunomia) Instance}
\usage{
injectCostData(connection)
}
\value{
Invisibly returns \code{TRUE} upon successful database insertion. Called for its side effect.
}
\description{
Inject Synthetic Payer Plan and Wide-Format Cost Data into an OMOP CDM (Eunomia) Instance
}
