% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CostCovariateData.R
\name{loadCostCovariateData}
\alias{loadCostCovariateData}
\title{Load CostCovariateData from file}
\usage{
loadCostCovariateData(file)
}
\arguments{
\item{file}{Path to RDS file}
}
\value{
CostCovariateData object
}
\description{
Load CostCovariateData from file
}
