% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CostCovariateData.R
\name{print.CostCovariateData}
\alias{print.CostCovariateData}
\title{Print CostCovariateData object}
\usage{
\method{print}{CostCovariateData}(x, ...)
}
\arguments{
\item{x}{CostCovariateData object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the input
}
\description{
Print CostCovariateData object
}
