% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CostCovariateSettings.R
\name{print.costCovariateSettings}
\alias{print.costCovariateSettings}
\title{Print Cost Covariate Settings}
\usage{
\method{print}{costCovariateSettings}(x, ...)
}
\arguments{
\item{x}{A costCovariateSettings object.}

\item{...}{Additional arguments (not used).}
}
\value{
Invisibly returns the input object.
}
\description{
Print Cost Covariate Settings
}
