% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CostCovariateData.R
\name{saveCostCovariateData}
\alias{saveCostCovariateData}
\title{Save CostCovariateData to a file}
\usage{
saveCostCovariateData(costCovariateData, file)
}
\arguments{
\item{costCovariateData}{A CostCovariateData object.}

\item{file}{The path to the file where the data should be saved.}
}
\description{
Saves the \code{CostCovariateData} object to a file. The Andromeda data is saved in a
zip file in the specified folder.
}
