% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizeCost.R
\name{standardizeCost}
\alias{standardizeCost}
\title{Standardize Costs for Inflation}
\usage{
standardizeCost(
  costCovariateData,
  inflationTable,
  connection,
  cdmDatabaseSchema
)
}
\arguments{
\item{costCovariateData}{A CostCovariateData object, requires person-level data.}

\item{inflationTable}{A data frame with 'year' and 'adjustment_factor' columns. The factor
should represent how to multiply a cost from that 'year' to get to the
standard year's value.}

\item{connection}{Database connection.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data resides.}
}
\value{
A new CostCovariateData object with adjusted costs.
}
\description{
Adjusts the cost values in a CostCovariateData object to a standard year's currency value
using an inflation adjustment table.
}
