% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CostCovariateData.R
\name{summary.CostCovariateData}
\alias{summary.CostCovariateData}
\title{Summary of CostCovariateData}
\usage{
\method{summary}{CostCovariateData}(object, ...)
}
\arguments{
\item{object}{A CostCovariateData object.}

\item{...}{Additional arguments (ignored).}
}
\value{
A tibble with the summary statistics.
}
\description{
Provides a summary of the covariates. This function executes a query on the
Andromeda object and returns an in-memory data frame.
}
